Change default max values within range of Duration.toMinutes method

+++ controller/src/main/java/io/pravega/controller/util/Config.java
@@ -178,11 +178,11 @@ public final class Config {
     public static final Property<RetentionType> PROPERTY_RETENTION_TYPE = Property.named(
             "retention.type", RetentionType.TIME, "retention.type");
 
-    public static final Property<Long> PROPERTY_RETENTION_MIN_VALUE = Property.named(
-            "retention.min.value", Long.MAX_VALUE, "retention.minVal");
+    public static final Property<Integer> PROPERTY_RETENTION_MIN_VALUE = Property.named(
+            "retention.min.value", Integer.MAX_VALUE, "retention.minVal");
 
-    public static final Property<Long> PROPERTY_RETENTION_MAX_VALUE = Property.named(
-            "retention.max.value", Long.MAX_VALUE, "retention.maxVal");
+    public static final Property<Integer> PROPERTY_RETENTION_MAX_VALUE = Property.named(
+            "retention.max.value", Integer.MAX_VALUE, "retention.maxVal");
 
     public static final Property<Boolean> PROPERTY_GLOBAL_RETENTION_POLICY = Property.named(
             "retention.global.policy.enabled", false, "retention.globalPolicy");
@@ -295,9 +295,9 @@ public final class Config {
     public static final RetentionType RETENTION_TYPE;
 
     public static final boolean GLOBAL_RETENTION_POLICY;
-    public static final long RETENTION_MIN_VALUE;
+    public static final int RETENTION_MIN_VALUE;
 
-    public static final long RETENTION_MAX_VALUE;
+    public static final int RETENTION_MAX_VALUE;
     public static final int RETENTION_THREAD_POOL_SIZE;
 
     // Watermarking Configuration
@@ -382,8 +382,8 @@ public final class Config {
         MINIMUM_RETENTION_FREQUENCY_IN_MINUTES = p.getInt(PROPERTY_RETENTION_FREQUENCY_MINUTES);
         RETENTION_BUCKET_COUNT = p.getInt(PROPERTY_RETENTION_BUCKET_COUNT);
         RETENTION_TYPE = p.getEnum(PROPERTY_RETENTION_TYPE, RetentionType.class);
-        RETENTION_MIN_VALUE = p.getLong(PROPERTY_RETENTION_MIN_VALUE);
-        RETENTION_MAX_VALUE = p.getLong(PROPERTY_RETENTION_MAX_VALUE);
+        RETENTION_MIN_VALUE = p.getInt(PROPERTY_RETENTION_MIN_VALUE);
+        RETENTION_MAX_VALUE = p.getInt(PROPERTY_RETENTION_MAX_VALUE);
         GLOBAL_RETENTION_POLICY = p.getBoolean(PROPERTY_GLOBAL_RETENTION_POLICY);
         RETENTION_THREAD_POOL_SIZE = p.getInt(PROPERTY_RETENTION_THREAD_COUNT);
         MINIMUM_WATERMARKING_FREQUENCY_IN_SECONDS = p.getInt(PROPERTY_WATERMARKING_FREQUENCY_SECONDS);
